+++ packages/asym-key/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@peculiar/asn1-rfc5958",
+  "name": "@peculiar/asn1-asym-key",
   "version": "2.3.4",
   "description": "Provides ASN.1 schema definitions for the asymmetric key packages format specified in RFC5958. This format is used to encapsulate private and public keys, along with additional parameters and metadata, for a variety of asymmetric key algorithms.",
   "keywords": [
