+++ packages/asym-key/test/test.spec.ts
@@ -1,5 +1,5 @@
 import * as assert from "node:assert";
-import { OneAsymmetricKey } from "@peculiar/asn1-rfc5958";
+import { OneAsymmetricKey } from "@peculiar/asn1-asym-key";
 import { AsnConvert } from "@peculiar/asn1-schema";
 import { id_rsaEncryption } from "@peculiar/asn1-rsa";
