+++ packages/x509-post-quantum/src/composite_keys.ts
@@ -1,4 +1,4 @@
-import { OneAsymmetricKey } from "@peculiar/asn1-rfc5958";
+import { OneAsymmetricKey } from "@peculiar/asn1-rfc8226";
 import { AsnArray, AsnType, AsnTypeTypes, BitString, OctetString } from "@peculiar/asn1-schema";
 import { AlgorithmIdentifier, SubjectPublicKeyInfo } from "@peculiar/asn1-x509";
